Add option to override peer data storage used with S130 and S132
This is required when using the (current version of the) bootloader
since the bootloader uses the legacy DM for accessing the peer data
instead of the PM. This causes the bootlader to fall back on the
"increment mac address" workaround which causes the button-less DFU
not to work properly.

See #3363.